requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langgraph>=0.0.20

# Database and data processing
pandas>=2.0.0

# Environment and utilities
python-dotenv>=1.0.0

# Optional for enhanced features
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0